[decoder] Remove subclasses in the header
This patch removes subclasses in the tensordec-boundingbox header.
Each box properties are located inside box_properties subdir.
