+++ ext/nnstreamer/tensor_decoder/box_properties/meson.build
@@ -1,0 +1,7 @@
+decoder_sub_bounding_boxes_sources += files(
+'mobilenetssd.cc',
+  'mobilenetssdpp.cc',
+  'ovdetection.cc',
+  'yolo.cc',
+  'mppalmdetection.cc'
+)
